Changing back the dalvik heapsize to 256M.
Runtime changes in dalvik parameters are not taking effect in the actual system.
They are just changing the dalvik numbers in getprop. So reverted the
runtime dalvik changes. As we can't differentiate the density during the compile time
setting the heapgrowth limit to max value among all the resolutions. (i.e, 128m)

Change-Id: Idaa6874ab006dfeffa60bd0bff7e66a530bc0658